+++ src/StimulusBundle/tests/AssetMapper/ControllersMapGeneratorTest.php
@@ -13,7 +13,6 @@ namespace Symfony\UX\StimulusBundle\Tests\AssetMapper;
 
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\AssetMapper\AssetMapperInterface;
-use Symfony\Component\AssetMapper\ImportMap\ImportMapConfigReader;
 use Symfony\Component\AssetMapper\MappedAsset;
 use Symfony\UX\StimulusBundle\AssetMapper\AutoImportLocator;
 use Symfony\UX\StimulusBundle\AssetMapper\ControllersMapGenerator;
@@ -54,17 +53,11 @@ class ControllersMapGeneratorTest extends TestCase
         $packageReader = new UxPackageReader(__DIR__.'/../fixtures');
 
         $autoImportLocator = $this->createMock(AutoImportLocator::class);
-        if (class_exists(ImportMapConfigReader::class)) {
-            $autoImportLocator->expects($this->any())
-                ->method('locateAutoImport')
-                ->willReturnCallback(function ($path) {
-                    return new MappedControllerAutoImport('/path/to'.$path, false);
-                });
-        } else {
-            // @legacy for AssetMapper 6.3
-            $autoImportLocator->expects($this->never())
-                ->method('locateAutoImport');
-        }
+        $autoImportLocator->expects($this->any())
+            ->method('locateAutoImport')
+            ->willReturnCallback(function ($path) {
+                return new MappedControllerAutoImport('/path/to'.$path, false);
+            });
 
         $generator = new ControllersMapGenerator(
             $mapper,
@@ -106,11 +99,7 @@ class ControllersMapGeneratorTest extends TestCase
         $this->assertSame('fake-vendor/ux-package1/package-controller-second.js', $controllerSecond->asset->logicalPath);
         // lazy from user's controller.json
         $this->assertTrue($controllerSecond->isLazy);
-        // @legacy: assert can be without the conditional for AssetMapper 6.4+
-        if (class_exists(ImportMapConfigReader::class)) {
-            // 4 auto imports from package.json
-            $this->assertCount(4, $controllerSecond->autoImports);
-        }
+        $this->assertCount(4, $controllerSecond->autoImports);
 
         $helloControllerFromPackage = $map['fake-vendor--ux-package2--hello-controller'];
         $this->assertSame('fake-vendor/ux-package2/package-hello-controller.js', $helloControllerFromPackage->asset->logicalPath);